Extract error creation/checking
