+++ src/lib/interface.ts
@@ -1,4 +1,5 @@
 import type { AllModules } from "~/modules/main/schema";
+import { isErrorMessage } from "./errors";
 import type { PromiseOrNot, UnionToIntersection } from "./utils";
 
 export type FeltHandlers = {
@@ -172,12 +173,4 @@ export function method<TKey extends keyof StandardMethods, R>(
       };
     });
   };
-}
-
-function isErrorMessage(
-  event: MessageEvent,
-): event is MessageEvent<{ __error__: string }> {
-  return (
-    event.data && typeof event.data === "object" && "__error__" in event.data
-  );
 }
